+++ CHANGES.md
@@ -8,6 +8,10 @@ Summary of the changes done in each version.
 
  - Explicitly disable HTTP caching, to always obtain a fresh response from ZAP.
 
+### New APIs
+
+ - WebSockets ("websocket").
+
 ## 1.5.0 (2017-11-30)
 
 ### Updated APIs